fold globalManager into PortalWrapper
