+++ src/components/PortalWrapper/PortalWrapper.utils.js
@@ -6,6 +6,7 @@
 // https://github.com/ReactTraining/react-router/blob/3d233bf0b6dd5bf68d9bac9c94273ae25646b207/packages/react-router/modules/matchPath.js
 import pathToRegexp from 'path-to-regexp'
 import isNil from 'lodash.isnil'
+import { createUniqueIDFactory } from '../../utilities/id'
 
 const patternCache = {}
 const cacheLimit = 10000
@@ -32,7 +33,7 @@ const compilePath = (pattern, options) => {
 /**
  * Public API for matching a URL pathname to a path pattern.
  */
-const matchPath = (pathname, options = {}, parent) => {
+export const matchPath = (pathname, options = {}, parent) => {
   if (typeof options === 'string') options = { path: options }
 
   const { path, exact = false, strict = false, sensitive = false } = options
@@ -60,4 +61,48 @@ const matchPath = (pathname, options = {}, parent) => {
   }
 }
 
-export default matchPath
+export const setupManager = (namespace = '') => {
+  const ns =
+    typeof namespace !== 'string' || !namespace.length
+      ? createUniqueIDFactory('BlueGlobalManager')()
+      : namespace
+
+  window[ns] = window[ns] || []
+
+  // Welcome aboard, Mr. Manager!
+  // Wow, I'm Mr. Manager!
+  // Well, manager… we we just say manager.
+  return {
+    data: window[ns],
+    add: add(window[ns]),
+    remove: remove(window[ns]),
+    first: first(window[ns]),
+    last: last(window[ns]),
+    max: max(window[ns]),
+  }
+}
+
+export const add = manager => item => {
+  manager.push(item)
+  return manager
+}
+
+export const remove = manager => item => {
+  const index = manager.indexOf(item)
+  if (index >= 0) {
+    manager.splice(index, 1)
+  }
+  return manager
+}
+
+export const first = manager => () => {
+  return manager[0]
+}
+
+export const last = manager => () => {
+  return manager[manager.length - 1]
+}
+
+export const max = manager => () => {
+  return Math.max.apply(Math, manager)
+}
